+++ CHANGELOG.md
@@ -3,13 +3,28 @@
 ## v0.6.0-dev
 
 * Enhancements
-  * [View] Add support to format_encoders that automatically encodes rendered templates. This means a "user.json" template only needs to return a map (or any structure encodable to JSON) and it will be automatically encoded to JSON by Phoenix
+  * [Controller] Support `put_view/2` to configure which view to use when rendering in the controller
+  * [Controller] Support templates as an atom in `Phoenix.Controller.render/3` as a way to explicitly render templates based on the request format
+  * [Controller] Split paths from external urls in `redirect/2`
+  * [Controller] `json/2` automatically encodes the data to JSON by using the registered `:format_encoders`
+  * [Controller] `html/2`, `json/2`, `text/2`, `redirect/2` and `render/3` now halt automatically
+  * [Controller] Add `accepts/2` for content negotiation
+  * [Controller] Add `put_layout_formats/2` and `layout_formats/1` to configure and read which formats have a layout when rendering
+  * [View] Assigns are always guaranteed to be maps
+  * [View] Add support to `format_encoders` that automatically encodes rendered templates. This means a "user.json" template only needs to return a map (or any structure encodable to JSON) and it will be automatically encoded to JSON by Phoenix
   * [View] Add a .exs template engine
 
+* Deprecations
+  * [Controller] `html/3`, `json/3`, `text/3` and `redirect/3` were deprecated in favor of using `put_status/2`
+  * [Controller] `redirect(conn, url)` was deprecated in favor of `redirect(conn, to: url)`
+
 * Backwards incompatible changes
+  * [Controller] Passing a string to render without format in the controller, as in `render(conn, "show")` no longer works. You should either make the format explicit `render(conn, "show.html")` or use an atom `render(conn, :show)` to dynamically render based on the format
+  * [View] Using `:within` was renamed in favor of `:layout` for rendering with layouts
   * [View] Your application should now directly use Phoenix.View in its main view and specify further configuration in the `using(...)` section
   * [View] Template engines now should implement compile and simply return the quoted expression of the function body instead of the quoted expression of the render function
-  * [Router] `PUT` route generation for the `:update` action has been dropped in favor of `PATCH`.
+  * [Router] `PUT` route generation for the `:update` action has been dropped in favor of `PATCH`
+  * [Router] Router no longer defines default :browser and :api pipelines
 
 * Bug fixes
   * [Router] Generate correct route for helper path on root