+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.6.0-dev
 
-See the [`0.5.x` to `0.6.0 upgrade instructions](https://gist.github.com/chrismccord/e774e6ab5220e6505a03) for upgrading your
+See the [`0.5.x` to `0.6.0` upgrade instructions](https://gist.github.com/chrismccord/e774e6ab5220e6505a03) for upgrading your
 existing applications.
 
 * Enhancements